dev_requirements.txt
coverage==4.0.3
pytest==2.8.5
mock==1.3.0
pytest-asyncio==0.3.0
pytest-catchlog==1.2.2
pyzmq==15.2.0
